moves files into root
